+++ eas-20210701/src/Models/CreateGatewayIntranetLinkedVpcPeerShrinkRequest.php
@@ -9,11 +9,15 @@ use AlibabaCloud\Tea\Model;
 class CreateGatewayIntranetLinkedVpcPeerShrinkRequest extends Model
 {
     /**
+     * @description The list of VPC peers.
+     *
      * @var string
      */
     public $peerVpcsShrink;
 
     /**
+     * @description The VPC ID. To obtain the VPC ID, see [ListGatewayIntranetLinkedVpc](https://help.aliyun.com/document_detail/2621223.html).
+     *
      * @example vpc-2zetuli9ws0qgjd******
      *
      * @var string