+++ eas-20210701/src/Models/DeleteGatewayIntranetLinkedVpcPeerResponseBody.php
@@ -9,6 +9,8 @@ use AlibabaCloud\Tea\Model;
 class DeleteGatewayIntranetLinkedVpcPeerResponseBody extends Model
 {
     /**
+     * @description The ID of the private gateway.
+     *
      * @example gw-1uhcqmsc7x22******
      *
      * @var string
@@ -16,6 +18,8 @@ class DeleteGatewayIntranetLinkedVpcPeerResponseBody extends Model
     public $gatewayId;
 
     /**
+     * @description The message that is returned.
+     *
      * @example Successfully delete intranet linked vpc Peer for gateway
      *
      * @var string
@@ -23,6 +27,8 @@ class DeleteGatewayIntranetLinkedVpcPeerResponseBody extends Model
     public $message;
 
     /**
+     * @description The ID of the request.
+     *
      * @example 40325405-579C-4D82****
      *
      * @var string