+++ eas-20210701/src/Models/DeleteGatewayIntranetLinkedVpcPeerShrinkRequest.php
@@ -9,11 +9,15 @@ use AlibabaCloud\Tea\Model;
 class DeleteGatewayIntranetLinkedVpcPeerShrinkRequest extends Model
 {
     /**
+     * @description The VPC peer.
+     *
      * @var string
      */
     public $peerVpcsShrink;
 
     /**
+     * @description The ID of the associated VPC. To obtain the VPC ID, see [ListGatewayIntranetLinkedVpc](https://help.aliyun.com/document_detail/2621223.html).
+     *
      * @example vpc-2zetuli9ws0qgjd******
      *
      * @var string