+++ eas-20210701/src/Models/ListGatewayIntranetSupportedZoneResponseBody.php
@@ -1,0 +1,63 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Eas\V20210701\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class ListGatewayIntranetSupportedZoneResponseBody extends Model
+{
+    /**
+     * @example 40325405-579C-4D82****
+     *
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var string[]
+     */
+    public $zones;
+    protected $_name = [
+        'requestId' => 'RequestId',
+        'zones'     => 'Zones',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+        if (null !== $this->zones) {
+            $res['Zones'] = $this->zones;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return ListGatewayIntranetSupportedZoneResponseBody
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+        if (isset($map['Zones'])) {
+            if (!empty($map['Zones'])) {
+                $model->zones = $map['Zones'];
+            }
+        }
+
+        return $model;
+    }
+}